Add files via upload
Pig is a simple dice game first described in print by John Scarne in 1945. Players take turns to roll a single die as many times as they wish, adding all roll results to a running total, but losing their gained score for the turn if they roll a 1. First to 100 wins. This project is written in JavaScript making use a DOM manipulation while also keeping track of scores on the back-end.